reorganized the files and passed test on simulator/real robot
