+++ real_robots/omnirobot_server.py
@@ -22,7 +22,7 @@ from tf.transformations import euler_from_quaternion
 
 from real_robots.constants import *
 from real_robots.utils import sendMatrix
-from real_robots.omnirobot_utils import OmnirobotManagerBase
+from real_robots.omnirobot_utils.omnirobot_manager_base import OmnirobotManagerBase
 assert USING_OMNIROBOT, "Please set USING_OMNIROBOT to True in real_robots/constants.py"
 
 NO_TARGET_MODE = False
@@ -348,7 +348,7 @@ class OmnirobotManager(OmnirobotManagerBase):
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description="Server for omnirobot")
-    parser.add_argument('--camera-info-path', default="real_robots/omnirobot_simulator_utils/cam_calib_info.yaml", type=str,
+    parser.add_argument('--camera-info-path', default="real_robots/omnirobot_utils/cam_calib_info.yaml", type=str,
                         help="camera calibration file generated by ros calibrate")
     args = parser.parse_args()
     with open(args.camera_info_path, 'r') as stream:
