--- real_robots/omnirobot_simulator_utils/__init__.py
@@ -1,3 +1,0 @@
-from .marker_finder import *
-
-from .marker_render import *
