+++ real_robots/omnirobot_utils/red_target_margin4_pixel_480x480.png
