+++ real_robots/omnirobot_utils/robot_margin3_pixel_only_tag.png
