+++ real_robots/omnirobot_utils/robot_margin4_pixel.png
