+++ real_robots/omnirobot_utils/target_margin4_pixel.png
